Move some warden functions to own file
